Natural regeneration of some commercial timber tree species following selective logging in a semi deciduous forest in the east region of Cameroon
A critical step in sustainable forest management is to ensure the establishment and regeneration of seedlings and sapling of exploitable tree species following logging. Since selective logging is one of the main silvicultural practices in Cameroon, a detailed understanding of regeneration following selective logging is vital. This study evaluated the natural regeneration of some commercial timber species in logged and unlogged forest types in two forest management units (FMU) in the East Region of Cameroon (FMU 10052 and 10025). Two transects of 5000 x 50m each were established in logged and unlogged forest types. Eleven commercial tree species were assessed for fruit fall, the number of seedlings established and the height increment of the established seedlings. Three of these commercial tree species fruited in both forest types. Fruit fall was significantly higher (p≤0.001)in the logged forest (492 fruits/ha)than in the unlogged forest (52 fruits/ha). Comparing species that fruited in both forest types Klainedoxa gabonensis recorded the highest number of fruit fall(84 fruits/ha) and least (0.24 fruits/ha) in the logged and unlogged forest types respectively. Seedling establishment was significantly higher (p≤0.001) in the unlogged forest (404 seedlings/ha) than in the logged forest (72 seedlings /ha). Seedling performance was better in the unlogged forest compared to the logged forest (low mortality rate). Due to the low seedling establishment and performance of these species in the logged forest, seed trees should be marked and protected prior and after logging as prescribed in the sustainable forest management.